Questions change properly now, Icons added, New Bug
Questions now change properly when submit button is first clicked on page load. 

Icons added to display if questions have been right or wrong. 

New bug with displaying score. Says unable to read correct_answer. I believe it is an index out of range error and the app is attempting to reach out for another question.